+++ R/bind_ard.R
@@ -7,10 +7,18 @@
 #'   ARDs to combine. Each argument can either be an ARD,
 #'   or a list of ARDs. Columns are matched by name, and any missing
 #'   columns will be filled with `NA`.
+#' @param .distinct (`logical`)\cr
+#'   logical indicating whether to remove non-distinct values from the ARD.
+#'   Duplicates are checked across grouping variables, primary variables,
+#'   context (if present), the **statistic name and the statistic value**.
+#'   Default is `FALSE`. If a statistic name and value is repeated and `.distinct=TRUE`,
+#'   the more recently added statistics will be retained, and the other(s) omitted.
 #' @param .update (`logical`)\cr
-#'   logical indicating whether to update duplicate ARD statistics.
-#'   Default is `FALSE`. If a statistic type is repeated and `.update=TRUE`,
-#'   the more recently added statistics will be retained, and the others omitted.
+#'   logical indicating whether to update ARD and remove duplicated named statistics.
+#'   Duplicates are checked across grouping variables, primary variables, and the
+#'   **statistic name**.
+#'   Default is `FALSE`. If a statistic name is repeated and `.update=TRUE`,
+#'   the more recently added statistics will be retained, and the other(s) omitted.
 #' @param .order (`logical`)\cr
 #'   logical indicating whether to order the rows of the stacked ARDs, allowing
 #'   statistics that share common group and variable values to appear in
@@ -26,10 +34,11 @@
 #' ard <- ard_categorical(ADSL, by = "ARM", variables = "AGEGR1")
 #'
 #' bind_ard(ard, ard, .update = TRUE)
-bind_ard <- function(..., .update = FALSE, .order = FALSE, .quiet = FALSE) {
+bind_ard <- function(..., .distinct = TRUE, .update = FALSE, .order = FALSE, .quiet = FALSE) {
   set_cli_abort_call()
 
   # check inputs ---------------------------------------------------------------
+  check_scalar_logical(.distinct)
   check_scalar_logical(.update)
   check_scalar_logical(.order)
   check_scalar_logical(.quiet)
@@ -37,13 +46,37 @@ bind_ard <- function(..., .update = FALSE, .order = FALSE, .quiet = FALSE) {
   # stack ARDs -----------------------------------------------------------------
   data <- dplyr::bind_rows(...)
 
-  # check for duplicates -------------------------------------------------------
+  # check for non-distinct statistics ------------------------------------------
+  not_distinct <-
+    dplyr::select(data, all_ard_groups(), all_ard_variables(), any_of("context"), "stat_name", "stat")[seq(nrow(data), 1L), ] |>
+    duplicated()
+  if (any(not_distinct) && isTRUE(.distinct)) {
+    if (isFALSE(.quiet)) {
+      cli::cli_inform(c(
+        "i" = "{sum(not_distinct)} row{?s} with {.emph duplicated statistic values} {?has/have} been removed.",
+        "*" = "See {.help [cards::bind_ard(.distinct)](cards::bind_ard)} for details."
+      ))
+    }
+    data <-
+      dplyr::filter(
+        data,
+        .by = c(all_ard_groups(), all_ard_variables(), "stat_name"),
+        dplyr::row_number() == dplyr::n()
+      )
+  }
+
+  # check for duplicate stat_name ----------------------------------------------
   dupes <-
     dplyr::select(data, all_ard_groups(), all_ard_variables(), "stat_name")[seq(nrow(data), 1L), ] |>
     duplicated()
 
   if (any(dupes) && isTRUE(.update)) {
-    if (isFALSE(.quiet)) cli::cli_inform(c("i" = "{sum(dupes)} duplicate observation{?/s} removed."))
+    if (isFALSE(.quiet)) {
+      cli::cli_inform(c(
+        "i" = "{sum(dupes)} row{?s} with {.emph duplicated statistic names} {?has/have} been removed.",
+        "*" = "See {.help [cards::bind_ard(.update)](cards::bind_ard)} for details."
+      ))
+    }
     data <-
       dplyr::filter(
         data,
@@ -51,7 +84,11 @@ bind_ard <- function(..., .update = FALSE, .order = FALSE, .quiet = FALSE) {
         dplyr::row_number() == dplyr::n()
       )
   } else if (any(dupes) && isFALSE(.update)) {
-    cli::cli_abort(c("!" = "{sum(dupes)} duplicate observation{?/s} found."),
+    cli::cli_abort(
+      c(
+        "!" = "{sum(dupes)} row{?s} with {.emph duplicated statistic names} {?has/have} been found.",
+        "i" = "See {.help [cards::bind_ard(.update)](cards::bind_ard)} for details."
+      ),
       call = get_cli_abort_call()
     )
   }